+++ test/test_utils.py
@@ -4,7 +4,7 @@ __email__ = 'info@3liz.org'
 
 import unittest
 
-from qgis_plugin_manager.utils import parse_version
+from qgis_plugin_manager.utils import parse_version, similar_names
 
 
 class TestUtils(unittest.TestCase):
@@ -14,3 +14,13 @@ class TestUtils(unittest.TestCase):
         self.assertListEqual([3, 10, 0], parse_version("3.10"))
         self.assertListEqual([3, 10, 0], parse_version("3.10.0"))
         self.assertIsNone(parse_version(""))
+
+    def test_similar_names(self):
+        """ Test about similar names in the XML file. """
+        self.assertListEqual(
+            ['Lizmap'],
+            similar_names(
+                'lizma',  # typo wanted
+                ['a', 'Lizmap', 'QuickOSM', ]
+            )
+        )